Extract `Timeout` from `DBTimeout`
