+++ richardPHJBensonBot/src/main/scala/com/benkio/richardphjbensonbot/DBTimeout.scala
@@ -2,50 +2,12 @@ package com.benkio.richardphjbensonbot
 
 import cats.effect.Async
 import cats.implicits._
+import com.benkio.richardphjbensonbot.model.Timeout
 import doobie.implicits._
 import doobie.implicits.javasql._
 import doobie.Transactor
 import doobie._
 import log.effect.LogWriter
-import telegramium.bots.Message
-
-import java.sql.Timestamp
-import java.time.Instant
-import scala.concurrent.duration._
-import scala.util.Try
-
-final case class Timeout(chat_id: Long, timeout_value: String, last_interaction: Timestamp)
-
-object Timeout {
-  def isExpired(timeout: Timeout): Boolean = {
-    val now = Timestamp.from(Instant.now())
-    now.after(
-      /*last interaction + timeout value*/
-      new Timestamp(timeout.last_interaction.getTime() + timeout.timeout_value.toInt)
-    )
-  }
-
-  def defaultTimeout(chatId: Long): Timeout = Timeout(
-    chat_id = chatId,
-    timeout_value = "0",
-    last_interaction = Timestamp.from(Instant.now())
-  )
-
-  def apply(m: Message, timeout: String): Option[Timeout] =
-    Try(
-      timeout
-        .split(":")
-        .map(_.toLong)
-        .zip(List(HOURS, MINUTES, SECONDS))
-        .map { case (value, timeUnit) =>
-          Duration(value, timeUnit)
-        }
-        .reduce(_ + _)
-    )
-      .map(duration => defaultTimeout(m.chat.id).copy(timeout_value = duration.toMillis.toString))
-      .toOption
-
-}
 
 sealed trait DBTimeout[F[_]] {
   def getOrDefault(chatId: Long): F[Timeout]
